moved resource interfaces to root
